+++ src/check-dependencies-depcheck-parser.js
@@ -1,0 +1,12 @@
+'use strict';
+
+const babylon = require('babylon');
+
+module.exports = function parser(content) {
+  return babylon.parse(content, {
+    sourceType: 'module',
+
+    // Enable all possible babylon plugins.
+    plugins: ['flow', 'jsx', '*']
+  });
+};