+++ src/jest-preprocessor.js
@@ -1,32 +1,14 @@
 'use strict';
 
-const babel = require('babel-core');
 const globals = require('./index-globals');
+const preprocessor = require('./lib/helper-preprocessor');
 
 module.exports = {
   process(src, filename) {
-    if (!babel.util.canCompile(filename)) {
-      // You might use a webpack loader in your project
-      // that allows you to load custom files (e.g. css, less, scss).
-      // Although it is working with webpack but jest
-      // don't know how to handle these files.
-      // You should never use these files unmocked
-      // otherwise you might encounter unexpected behavior.
-      // Changing the content of these files
-      // to return the raw content.
-      return `module.exports = ${JSON.stringify(src)}`;
-    }
+    const presets = [require.resolve('babel-preset-jest')].concat(globals.babel.presets);
 
-    if (filename.indexOf('node_modules') === -1) {
-      const presets = [require.resolve('babel-preset-jest')].concat(globals.babel.presets);
-
-      return babel.transform(src, Object.assign({}, globals.babel, {
-        filename,
-        retainLines: true,
-        presets
-      })).code;
-    }
-
-    return src;
+    return preprocessor.processWithBabel(src, filename, Object.assign({}, globals.babel, {
+      presets
+    }));
   }
 };